# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simple Invoice API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://invoice-app.test/api/v1",
            "description": "Development server"
        }
    ],
    "paths": {
        "/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get the list of addresses",
                "operationId": "getAddressList",
                "responses": {
                    "200": {
                        "description": "The addresses"
                    }
                }
            }
        },
        "/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get the list of clients (paginated)",
                "operationId": "getClientList",
                "responses": {
                    "200": {
                        "description": "The clients"
                    }
                }
            }
        },
        "/clients/{clientId}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get a client by id",
                "operationId": "getClient",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "required": true,
                        "description": "The id of the client",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The client"
                    }
                }
            }
        },
        "/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Get the list of invoices (paginated)",
                "operationId": "getInvoiceList",
                "responses": {
                    "200": {
                        "description": "The invoices"
                    }
                }
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Create a new invoice",
                "operationId": "createInvoice",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "invoiceDate",
                                    "paymentTermId",
                                    "clientId",
                                    "addressId",
                                    "items"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Invoice #1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is the first invoice"
                                    },
                                    "invoiceDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-10-01"
                                    },
                                    "paymentTermId": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "clientId": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "addressId": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "draft"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Item #1"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": "100.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice created"
                    },
                    "500": {
                        "description": "Invoice not created"
                    }
                }
            }
        },
        "/invoices/{invoiceId}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Get an invoice by id",
                "operationId": "getInvoice",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "required": true,
                        "description": "The id of the invoice",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The invoice"
                    }
                }
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Update an invoice",
                "operationId": "updateInvoice",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "required": true,
                        "description": "The id of the invoice",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "invoiceDate",
                                    "paymentTermId",
                                    "addressId",
                                    "items"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Invoice #1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is the first invoice"
                                    },
                                    "invoiceDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-10-01"
                                    },
                                    "paymentTermId": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "addressId": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "(int)1 or null for new items"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Item #1"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": "100.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice updated"
                    },
                    "500": {
                        "description": "Invoice not updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Delete an invoice",
                "operationId": "deleteInvoice",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "required": true,
                        "description": "The id of the invoice",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice deleted"
                    },
                    "500": {
                        "description": "Invoice not deleted"
                    }
                }
            }
        },
        "/my/profile": {
            "get": {
                "tags": [
                    "My"
                ],
                "summary": "Get the user profile",
                "operationId": "getMyProfile",
                "responses": {
                    "200": {
                        "description": "Profile"
                    }
                }
            }
        },
        "/my/settings": {
            "get": {
                "tags": [
                    "My"
                ],
                "summary": "Get the user settings",
                "operationId": "getMySettings",
                "responses": {
                    "200": {
                        "description": "Settings"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Addresses",
            "description": "Addresses"
        },
        {
            "name": "Clients",
            "description": "Clients"
        },
        {
            "name": "Invoices",
            "description": "Invoices"
        },
        {
            "name": "My",
            "description": "My"
        }
    ]
}